refactor(frontend): extract ResultSection into sub-component
Move ResultSection, renderToolResult, and extractFailureMessage out of
ToolExecutionCardView into a dedicated ResultSection sub-component
under components/, following the one-component-per-file convention.
