+++ apps/frontend/src/pages/chat/components/MessageList/components/ToolExecutionCard/components/ResultSection/index.ts
@@ -1,0 +1,1 @@
+export {ResultSectionView as ResultSection} from './ResultSectionView.js';
